Lottery and `insertPrefix` were added to main class. While all the "Utility" based methods were moved to `Utils`.
